pyproject.toml
[tool.poetry]
name = "airbnb-clone-backend"
version = "0.1.0"
description = ""
authors = ["lee seung ah <lee88067@gmail.com>"]
readme = "README.md"
packages = [{include = "airbnb_clone_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
pillow = "^9.4.0"
beautifulsoup4 = "^4.11.2"
strawberry-graphql = {extras = ["debug-server"], version = "^0.171.0"}
strawberry-graphql-django = "^0.9.4"
jwt = "^1.3.1"
pyjwt = "^2.6.0"
django-environ = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
